New data type: tagged string
In some cases the string data type is not
expressive enough. It requires key-specific
handling in the application, that is,
the key name becomes hardcoded if the string
has any special meaning to the app.

By introducing a new data type we enforce
a convention where a schema can define
a "special" kind (in fact, multiple kinds)
of string and it will be expressed in the
generated value, which will be a tagged tuple.

This approach avoids parser extensions,
whether this means native syntax or
a complex data type extension such as
the duration data type.

Closes #40.